+++ demo/src/main/java/com/loremipsum/demo/Follow.java
@@ -2,22 +2,22 @@ package com.loremipsum.demo;
 
 public class Follow {
     //instances
-    private Long followerId;
-    private Long followedId;
+    private int followerId;
+    private int followedId;
 
-    public Long getFollowerId(){
+    public int getFollowerId(){
         return this.followerId;
     }
 
-    public Long getFollowedId(){
+    public int getFollowedId(){
         return this.followedId;
     }
 
-    public void setFollowerId(Long newId){
+    public void setFollowerId(int newId){
         this.followerId = newId;
     }
 
-    public void setFollowedId(Long newId){
+    public void setFollowedId(int newId){
         this.followedId = newId;
     }
 }
